creating models/structs.go in which all the structs definition will be, and creating add_user_validation/validation.go
